Refactor: Form into Roles Component
